docs(Filter): move render trigger controls
